Improve test replayability
It took me a while, but I realised that using the saved `Rand.State`
was a much better way of replaying a test case than the random seed.

For example, the random seed will replay the non-reduced failing test
case, which is not the one we want - we always want the smallest
possible failing test case.

The current implementation base-64 encodes a gzipped state, which
works, but might not be optimal. I might rework this later.