Introduce the notion of test plan
This is more intended for future experiments than anything - in the
current version, there probably isn't a need for more than one test
plan.

The motivation is my having stumbled on this paper:
https://people.irisa.fr/Thomas.Genet/ACF/Bibliography/Isabelle/Advanced/exhaustiveQuickcheck.pdf

And now I want the ability to easily write a plan for this and plug it
in without having to fight against the library's design.