Improve growing search by growing an upper bound on random ints
Growing the size linearly gives OK results - provided the underlying
generators use it.

It dawned on me that, since I fully control random generation and the
convention is for smaller random numbers to lead to smaller generated
values, I could also set an upper bound on random values, and increase
that on each iteration.

I've gone for a geometric increase, so that we stay in the realm of
quite small values for a while, and then essentially drop the
constraint to explore the entire space.